+++ server/controllers/messaging.js
@@ -1,43 +1,6 @@
-import fs from "fs";
-import path from "path";
-
-import admin from "firebase-admin";
-
 import { queryPromise } from "../db/database.js";
 import HttpResponseWrapper from "../global/HttpResponseWrapper.js";
 import { addErrorTitle } from "../global/Logger.js";
-
-/* Initialize firebase */
-fs.readFile(
-  path.resolve(fs.realpathSync("."), "./files/serviceAccountKey.json"),
-  "utf-8",
-  (err, data) => {
-    if (err) {
-      console.error("Can't read service account file.", err);
-      return;
-    }
-    admin.initializeApp({
-      credential: admin.credential.cert(JSON.parse(data)),
-    });
-  }
-);
-
-/**
- *
- * @api {get} /messaging
- * @apiName GetUsersRegistratedPushUsers
- *
- * @apiSuccess (200) {array} List of users token
- *
- * @apiUse ErrorServer
- */
-function getUsersRegistratedPush(req, _res) {
-  const res = new HttpResponseWrapper(_res);
-
-  getAllUsersRegistrated()
-    .then((data) => res.sendResponse(200, data))
-    .catch(res.sendServerError);
-}
 
 /**
  *
@@ -52,7 +15,7 @@ function getUsersRegistratedPush(req, _res) {
  * @apiParamExample {object} Request-Example:
  * {
  *   "user": "nhoun",
- *   "messagingToken": "cjFl5JjYiaUdFfxuKefa...iv9ViE6_m"
+ *   "messagingToken": "cjFl5JjYiaUdFfxuKefa...iv9ViE6_m",
  * }
  *
  *
@@ -86,6 +49,25 @@ function updateToken(req, _res) {
     .catch(res.sendServerError);
 }
 
+/**
+ *
+ * @api {put} /messaging/token/remove
+ * @apiName RemoveUserToken
+ * @apiGroup Messaging
+ *
+ * @apiSuccess (200) {object} New messaging token
+ *
+ * @apiParamExample {object} Request-Example:
+ * {
+ *   "user": "nhoun",
+ * }
+ *
+ *
+ * @apiUse ErrorServer
+ * @apiError (400) InvalidUser The user is not valid
+ * @apiError (400) InvalidToken The token is not valid
+ * @apiError (404) NotFound User not found
+ */
 function removeToken(req, _res) {
   const res = new HttpResponseWrapper(_res);
 
@@ -108,19 +90,7 @@ function removeToken(req, _res) {
     .catch(res.sendServerError);
 }
 
-function sendNotificationToOneDevice(targetToken, data) {
-  const message = {
-    data: data,
-    token: targetToken,
-  };
-
-  admin
-    .messaging()
-    .send(message)
-    .catch((err) => console.error("Can't send notification to a single device", err));
-}
-
-export default { updateToken, removeToken, getUsersRegistratedPush, sendNotificationToOneDevice };
+export default { updateToken, removeToken };
 
 // ***** INTERNAL FUNCTIONS *****
 
@@ -139,16 +109,5 @@ function saveMessagingTokenInDatabase(user, messagingToken = undefined) {
         }
       })
       .catch((error) => reject(addErrorTitle(error, "Can't update messaging token", true)));
-  });
-}
-
-function getAllUsersRegistrated() {
-  return new Promise((resolve, reject) => {
-    const sql = `SELECT us_messaging_token
-                 FROM user
-                 WHERE us_messaging_token IS NOT NULL`;
-    queryPromise(sql)
-      .then((res) => resolve(res.map((value) => value.us_messaging_token)))
-      .catch((error) => reject(addErrorTitle(error, "Can't get all registrated push users", true)));
   });
 }
